pyproject.toml
[project]
name = "course-recommendation-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn==0.30.6",
  "google-cloud-aiplatform==1.67.1",
  "google-genai==0.5.0",
  "vertexai==1.67.1",
]
